:recycle: [refactor] 스테이지 play 카운트다운, 노래 재생 위젯화
